% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doppelgangR-package.R
\docType{package}
\name{doppelgangR-package}
\alias{doppelgangR-package}
\title{Identify likely duplicate samples from genomic or meta-data}
\description{
The main function is doppelgangR(), which takes as minimal input a list of
ExpressionSet object, and searches all list pairs for duplicated samples.
The search is based on the genomic data (exprs(eset)), phenotype/clinical
data (pData(eset)), and "smoking guns" - supposedly unique identifiers found
in pData(eset).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lwaldron/doppelgangR}
  \item \url{https://waldronlab.github.io/doppelgangR}
  \item Report bugs at \url{https://github.com/lwaldron/doppelgangR/issues}
}

}
\author{
Levi Waldron, Markus Riester, Marcel Ramos
}
\keyword{package}
