% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoppelGang-class.R
\docType{methods}
\name{print-methods}
\alias{print-methods}
\alias{print,DoppelGang-method}
\title{Print a DoppelGang object}
\usage{
\S4method{print}{DoppelGang}(x)
}
\description{
Prints the @summaryresults slot of a DoppelGang object.
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"DoppelGang\")")}{ Prints the @summaryresults
slot of a DoppelGang object. }

}
}

\keyword{methods}
