% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colFinder.R
\name{colFinder}
\alias{colFinder}
\title{Calculate pairwise similarities of colData between samples for a list
containing two DataFrame}
\usage{
colFinder(summex.list, ...)
}
\arguments{
\item{summex.list}{input: a list of DataFrame with two elements, or a
DataFrame. If the two elements are identical, return the correlation
matrix for pairs of samples in the first element.  If not identical, return
pairs between the two elements.}

\item{\dots}{Extra arguments passed on to colFinder}
}
\value{
A matrix of similarities between the colData of pairs of samples.
}
\description{
This function acts as a wrapper to colData to handle cases of one
DataFrame, a list of two identical DataFrame, or a list of two
different DataFrame
}
\author{
Fabio Da Col, Marcel Ramos
}
