% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoppelGang-class.R
\docType{class}
\name{DoppelGang-class}
\alias{DoppelGang-class}
\alias{DoppelGang}
\title{DoppelGang S4 class}
\description{
S4 class containing results of doppelgangR() function.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new(DoppelGang ...)}
}

\seealso{
\code{\link{plot,DoppelGang-method}}
}
\author{
Levi Waldron and Markus Riester
}
