% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoppelGang-class.R
\docType{class}
\name{DoppelGang-class}
\alias{DoppelGang-class}
\alias{DoppelGang}
\alias{DoppelGang-class}
\alias{summary,DoppelGang-method}
\alias{DoppelGang-class}
\alias{show,DoppelGang-method}
\alias{DoppelGang-class}
\alias{print,DoppelGang-method}
\title{DoppelGang S4 class}
\usage{
\S4method{summary}{DoppelGang}(object)

\S4method{show}{DoppelGang}(object)

\S4method{print}{DoppelGang}(x)
}
\arguments{
\item{x, object}{A DoppelGang class object}
}
\description{
S4 class containing results of doppelgangR() function.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new(DoppelGang ...)}
}

\seealso{
\code{\link{plot,DoppelGang-method}}
}
\author{
Levi Waldron and Markus Riester
}
