\name{doppelgangR-package}
\alias{doppelgangR-package}
\docType{package}
\title{Identify likely duplicate samples from genomic or meta-data}

\description{The main function is doppelgangR(), which takes as minimal
input a list of ExpressionSet object, and searches all list pairs for
duplicated samples.  The search is based on the genomic data (exprs(eset)),
phenotype/clinical data (pData(eset)), and "smoking guns" - supposedly unique
identifiers found in pData(eset).}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{doppelgangR}
\packageIndices{doppelgangR}
}
\author{Levi Waldron, Markus Riester, Marcel Ramos

Maintainer: \packageMaintainer{doppelgangR}
}

\keyword{ package }


