% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoppelGang-class.R
\docType{methods}
\name{summary-methods}
\alias{summary-methods}
\alias{summary,DoppelGang-method}
\title{Summarizes a DoppelGang object}
\usage{
\S4method{summary}{DoppelGang}(object)
}
\description{
Summarizes the @summaryresults slot of a DoppelGang object.
}
\section{Methods}{
 \describe{ \item{list("signature(object =
\"DoppelGang\")")}{ Summarizes the @summaryresults slot of a DoppelGang
object.  } }
}

\keyword{methods}
